Extract ORM-specific stuff only to adapters
I'm going to try to extract adapters into separate gems. This is the first
step - all ORM-related stuff goes to ORM adapters classes. The only exceptions
right now are inclusions of Sorcery to model class. This will be changed
as soon as adapters are extracted to other gems.

Currently tests are not passing due to the way modules are included in
lib/sorcery/model.rb, however when only one adapter is included, the tests
mostly pass (except for Authentication class tests)