Extract Mongoid adapter
This is experimental step in extracting adapters - I threw out loading ORMs code
so that gem works correctly, and extracted Mongoid adapter to separate gem.

Some changes were made in adapters' interface, which are not reflected in other
adapters yet. Tests are suppose to be failing as it's just POC and I'll work on
them later